diags/diag_disk: Fix build warnings
To fix strncpy() warnings, replace it with memcpy() and explicitly
handle null termination.

Fix snprintf() warnings by checking return value of snprintf(),
which returns negative value on error and returns size of format string
after substitution. If the size of format string is greater than or equal
to size of destination buffer, then this implies that the string was too
big to fit in the destination buffer. In either of the above cases,
snprintf() will fail. To indicate failure, display error message.

  CC       diags/diag_disk.o
In file included from /usr/include/string.h:519,
                 from diags/diag_disk.c:33:
In function 'strncpy',
    inlined from 'get_system_vpd' at diags/diag_disk.c:261:2,
    inlined from 'diag_disk' at diags/diag_disk.c:433:8:
/usr/include/bits/string_fortified.h:95:10: warning: '__builtin_strncpy' output may be truncated copying 8 bytes from a string of length 15 [-Wstringop-truncation]
   95 |   return __builtin___strncpy_chk (__dest, __src, __len,
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   96 |                                   __glibc_objsize (__dest));
      |                                   ~~~~~~~~~~~~~~~~~~~~~~~~~
In function 'strncpy',
    inlined from 'get_system_vpd' at diags/diag_disk.c:277:2,
    inlined from 'diag_disk' at diags/diag_disk.c:433:8:
/usr/include/bits/string_fortified.h:95:10: warning: '__builtin_strncpy' output may be truncated copying 9 bytes from a string of length 15 [-Wstringop-truncation]
   95 |   return __builtin___strncpy_chk (__dest, __src, __len,
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   96 |                                   __glibc_objsize (__dest));
      |                                   ~~~~~~~~~~~~~~~~~~~~~~~~~
diags/diag_disk.c: In function 'diag_disk':
diags/diag_disk.c:403:62: warning: '%s' directive output may be truncated writing up to 4352 bytes into a region of size 4065 [-Wformat-truncation=]
  403 |                 snprintf(filename, sizeof(filename) - 1, %s/%s,
      |                                                              ^~
In file included from /usr/include/stdio.h:888,
                 from diags/diag_disk.c:31:
/usr/include/bits/stdio2.h:71:10: note: '__builtin___snprintf_chk' output between 31 and 4383 bytes into a destination of size 4095
   71 |   return __builtin___snprintf_chk (__s, __n, __USE_FORTIFY_LEVEL - 1,
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   72 |                                    __glibc_objsize (__s), __fmt,
      |                                    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   73 |                                    __va_arg_pack ());
      |                                    ~~~~~~~~~~~~~~~~~
diags/diag_disk.c:191:54: warning: '%s' directive output may be truncated writing up to 4095 bytes into a region of size 4065 [-Wformat-truncation=]
  191 |         snprintf(filename, sizeof(filename) - 1, %s/%s,
      |                                                      ^~
......
  459 |         ret = open_output_xml_file(xml_filename);
      |                                    ~~~~~~~~~~~~
In file included from /usr/include/stdio.h:888,
                 from diags/diag_disk.c:31:
/usr/include/bits/stdio2.h:71:10: note: '__builtin___snprintf_chk' output between 31 and 4126 bytes into a destination of size 4095
   71 |   return __builtin___snprintf_chk (__s, __n, __USE_FORTIFY_LEVEL - 1,
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   72 |                                    __glibc_objsize (__s), __fmt,
      |                                    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   73 |                                    __va_arg_pack ());
      |                                    ~~~~~~~~~~~~~~~~~

+++ diags/diag_disk.c
@@ -188,8 +188,13 @@ static int open_output_xml_file(const char *xml_filename)
 	}
 
 
-	snprintf(filename, sizeof(filename) - 1, "%s/%s",
-		 DISK_OUTPUT_PATH, xml_filename);
+	rc = snprintf(filename, sizeof(filename) - 1, "%s/%s",
+			DISK_OUTPUT_PATH, xml_filename);
+	if (rc < 0 || (rc >= sizeof(filename) - 1)) {
+		fprintf(stderr, "%s:%d - Unable to format %s\n",
+				__func__, __LINE__, xml_filename);
+		return -1;
+	}
 
 	result_file = fopen(filename, "w");
 	if (!result_file)
@@ -258,7 +263,8 @@ static int get_system_vpd(char *machine_serial,
 	 */
 	if (strlen(serial) > 5)
 		start_index = strlen(serial) - 5;
-	strncpy(machine_serial, serial + start_index, SERIAL_NUM_LEN);
+	memcpy(machine_serial, serial + start_index, SERIAL_NUM_LEN);
+	machine_serial[SERIAL_NUM_LEN - 1] = '\0';
 
 	device_fd = open(DEVICE_TREE_MODEL, O_RDONLY);
 	if (device_fd < 0)
@@ -274,7 +280,8 @@ static int get_system_vpd(char *machine_serial,
 	if (strchr(model, ',') != NULL)
 		temp = strchr(model, ',') + 1;
 
-	strncpy(machine_type, temp, MACHINE_MODEL_LEN + 1);
+	memcpy(machine_type, temp, MACHINE_MODEL_LEN + 1);
+	machine_type[MACHINE_MODEL_LEN] = '\0';
 	*machine_model = strchr(machine_type, '-');
 	if (*machine_model == NULL) /* Failed to get model name */
 		return -1;
@@ -391,6 +398,7 @@ static int remove_old_log_file(void)
 	DIR *d;
 	struct dirent *namelist;
 	char filename[PATH_MAX];
+	int rc;
 
 	d = opendir(DISK_OUTPUT_PATH);
 	if (!d)
@@ -400,8 +408,14 @@ static int remove_old_log_file(void)
 		if (namelist->d_name[0] == '.')
 			continue;
 
-		snprintf(filename, sizeof(filename) - 1, "%s/%s",
+		rc = snprintf(filename, sizeof(filename) - 1, "%s/%s",
 				DISK_OUTPUT_PATH, namelist->d_name);
+		if (rc < 0 || rc >= (sizeof(filename) - 1)) {
+			fprintf(stderr, "%s:%d - Unable to format %s\n",
+				__func__, __LINE__, namelist->d_name);
+			continue;
+		}
+
 		if (unlink(filename) < 0) {
 			fprintf(stderr,
 			"\nUnable to remove old log file[%s]. continuing.\n\n",
